Moving some methods to search utilities
